+++ cmd/influxd/launcher/storage_test.go
@@ -10,13 +10,14 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/influxdata/influxdb"
+	"github.com/influxdata/influxdb/cmd/influxd/launcher"
 	"github.com/influxdata/influxdb/http"
 	"github.com/influxdata/influxdb/toml"
 	"github.com/influxdata/influxdb/tsdb/tsm1"
 )
 
 func TestStorage_WriteAndQuery(t *testing.T) {
-	l := RunLauncherOrFail(t, ctx)
+	l := launcher.RunTestLauncherOrFail(t, ctx)
 
 	org1 := l.OnBoardOrFail(t, &influxdb.OnboardingRequest{
 		User:     "USER-1",
@@ -53,7 +54,7 @@ func TestStorage_WriteAndQuery(t *testing.T) {
 }
 
 func TestLauncher_WriteAndQuery(t *testing.T) {
-	l := RunLauncherOrFail(t, ctx)
+	l := launcher.RunTestLauncherOrFail(t, ctx)
 	l.SetupOrFail(t)
 	defer l.ShutdownOrFail(t, ctx)
 
@@ -91,7 +92,7 @@ func TestLauncher_WriteAndQuery(t *testing.T) {
 }
 
 func TestLauncher_BucketDelete(t *testing.T) {
-	l := RunLauncherOrFail(t, ctx)
+	l := launcher.RunTestLauncherOrFail(t, ctx)
 	l.SetupOrFail(t)
 	defer l.ShutdownOrFail(t, ctx)
 
@@ -157,7 +158,7 @@ func TestLauncher_BucketDelete(t *testing.T) {
 }
 
 func TestStorage_CacheSnapshot_Size(t *testing.T) {
-	l := NewLauncher()
+	l := launcher.NewTestLauncher()
 	l.StorageConfig.Engine.Cache.SnapshotMemorySize = 10
 	l.StorageConfig.Engine.Cache.SnapshotAgeDuration = toml.Duration(time.Hour)
 	defer l.ShutdownOrFail(t, ctx)
@@ -203,7 +204,7 @@ func TestStorage_CacheSnapshot_Size(t *testing.T) {
 }
 
 func TestStorage_CacheSnapshot_Age(t *testing.T) {
-	l := NewLauncher()
+	l := launcher.NewTestLauncher()
 	l.StorageConfig.Engine.Cache.SnapshotAgeDuration = toml.Duration(time.Second)
 	defer l.ShutdownOrFail(t, ctx)
 
@@ -245,39 +246,4 @@ func TestStorage_CacheSnapshot_Age(t *testing.T) {
 	if got, exp := summary.Total, uint64(5); got != exp {
 		t.Fatalf("got %d series in TSM files, expected %d", got, exp)
 	}
-}
-
-// WriteOrFail attempts a write to the organization and bucket identified by to or fails if there is an error.
-func (l *Launcher) WriteOrFail(tb testing.TB, to *influxdb.OnboardingResults, data string) {
-	tb.Helper()
-	resp, err := nethttp.DefaultClient.Do(l.NewHTTPRequestOrFail(tb, "POST", fmt.Sprintf("/api/v2/write?org=%s&bucket=%s", to.Org.ID, to.Bucket.ID), to.Auth.Token, data))
-	if err != nil {
-		tb.Fatal(err)
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		tb.Fatal(err)
-	}
-
-	if err := resp.Body.Close(); err != nil {
-		tb.Fatal(err)
-	}
-
-	if resp.StatusCode != nethttp.StatusNoContent {
-		tb.Fatalf("unexpected status code: %d, body: %s, headers: %v", resp.StatusCode, body, resp.Header)
-	}
-}
-
-// FluxQueryOrFail performs a query to the specified organization and returns the results
-// or fails if there is an error.
-func (l *Launcher) FluxQueryOrFail(tb testing.TB, org *influxdb.Organization, token string, query string) string {
-	tb.Helper()
-
-	b, err := http.SimpleQuery(l.URL(), query, org.Name, token)
-	if err != nil {
-		tb.Fatal(err)
-	}
-
-	return string(b)
 }