+++ ui/src/clockface/components/context_menu/ContextMenu.tsx
@@ -17,20 +17,15 @@ import {
 
 import {ErrorHandling} from 'src/shared/decorators/errors'
 
-interface PassedProps {
+interface Props {
   children: JSX.Element | JSX.Element[]
   icon: IconFont
   onBoostZIndex?: (boostZIndex: boolean) => void
+  text: string
+  color: ComponentColor
+  shape: ButtonShape
+  testID: string
 }
-
-interface DefaultProps {
-  text?: string
-  color?: ComponentColor
-  shape?: ButtonShape
-  testID?: string
-}
-
-type Props = PassedProps & DefaultProps
 
 interface State {
   isExpanded: boolean
@@ -38,7 +33,7 @@ interface State {
 
 @ErrorHandling
 class ContextMenu extends Component<Props, State> {
-  public static defaultProps: DefaultProps = {
+  public static defaultProps = {
     color: ComponentColor.Primary,
     shape: ButtonShape.Square,
     text: '',