+++ ui/src/index.tsx
@@ -54,6 +54,7 @@ import OrgTemplatesIndex from 'src/organizations/containers/OrgTemplatesIndex'
 import TemplateExportOverlay from 'src/templates/components/TemplateExportOverlay'
 import TemplateImportOverlay from 'src/templates/components/TemplateImportOverlay'
 import CreateOrgOverlay from './organizations/components/CreateOrgOverlay'
+import VariableImportOverlay from 'src/variables/components/VariableImportOverlay'
 
 import OnboardingWizardPage from 'src/onboarding/containers/OnboardingWizardPage'
 
@@ -169,6 +170,10 @@ class Root extends PureComponent {
                               path="variables"
                               component={OrgVariablesIndex}
                             >
+                              <Route
+                                path="import"
+                                component={VariableImportOverlay}
+                              />
                               <Route
                                 path=":id/export"
                                 component={OrgVariableExportOverlay}